Removes IProcess, moves Scale to Quantity and optimizes some MultiMap functions
