--- src/main/kotlin/org/ageseries/libage/sim/IProcess.kt
@@ -1,12 +1,0 @@
-package org.ageseries.libage.sim
-
-/**
- * An interface that represents a process.
- */
-interface IProcess {
-    /**
-     * Runs the simulation for a specified time delta.
-     * @param dt The duration of time to simulate.
-     */
-    fun process(dt: Double)
-}
