--- src/test/kotlin/org/eln2/libelectric/sim/electrical/mna/IProcessTest.kt
@@ -1,19 +1,0 @@
-package org.eln2.libelectric.sim.electrical.mna
-
-import org.ageseries.libage.sim.IProcess
-import org.junit.jupiter.api.Assertions
-import org.junit.jupiter.api.Test
-
-internal class IProcessTest : IProcess {
-    private var timeRun: Double = 0.0
-
-    @Test
-    fun iProcessTest() {
-        repeat(20) { process(0.05); } // Run the simulation 20 times for 50 ms each.
-        Assertions.assertEquals(timeRun, 1.0, 0.001) // Total time elapsed should be one second.
-    }
-
-    override fun process(dt: Double) {
-        timeRun += dt
-    }
-}
